requirements.txt
gymnasium==0.29.1
numpy==1.26.4
stable-baselines3==2.3.2
torch==2.2.2
tqdm==4.66.4
matplotlib==3.8.4
pyyaml==6.0.2
pandas==2.2.2
minigrid==2.3.1
pygame==2.5.2
imageio==2.34.0
rich==13.9.4
